+++ resync/resource_list.py
@@ -1,16 +1,16 @@
 """ResourceSync Resource List object.
 
 A Resource List is a set of resources with some metadata for
-each resource. Comparison of resource lists from a source and
-a destination allows understanding of whether the two are in
-sync or whether some resources need to be updated at the
-destination.
+each resource. Every resource must have a uri attribute.
+Comparison of resource lists from a source and a destination
+allows understanding of whether the two are in sync or whether
+some resources need to be updated at the destination.
 
 There may also be metadata about the Resource List, and links
 to other ResourceSync documents. Metadata include the
 timestamp of the Resource List (md_at) and, optionally, the
 timestamp when creation of the Resource List was completed
-(md_completed).at the top level. These include a creation timestamp
+(md_completed) at the top level. These include a creation timestamp
 (from) and links to the Capability List.
 
 Described in specification at:
@@ -18,6 +18,7 @@ http://www.openarchives.org/rs/resourcesync#DescResources
 """
 
 import collections.abc
+from collections import OrderedDict
 import os
 from datetime import datetime
 import re
@@ -29,29 +30,62 @@ from .mapper import Mapper, MapperError
 from .url_authority import UrlAuthority
 
 
-class ResourceListDict(dict):
+class SortedIterMixin(object):
+    """Mixin to provide sorted_iter() method."""
+
+    def sorted_iter(self):
+        """Iterator over all the resources in this dict by sorted key order."""
+        self._sorted_iter_next_list = sorted(self.keys(), reverse=True)
+        return iter(self._sorted_iter_next, None)
+
+    def _sorted_iter_next(self):
+        if (len(self._sorted_iter_next_list) > 0):
+            return self[self._sorted_iter_next_list.pop()]
+        else:
+            return None
+
+
+class ResourceListDict(dict, SortedIterMixin):
     """Default implementation of class to store resources in ResourceList.
 
     Key properties of this class are:
     - has add(resource) method
-    - is iterable and results given in alphanumeric order by resource.uri
+    - is iterable and gives resources (not keys) in alphanumeric order by
+      resource.uri
     """
 
     def __iter__(self):
         """Iterator over all the resources in this ResourceListDict."""
-        self._iter_next_list = sorted(self.keys())
-        self._iter_next_list.reverse()
-        return(iter(self._iter_next, None))
-
-    def _iter_next(self):
-        if (len(self._iter_next_list) > 0):
-            return(self[self._iter_next_list.pop()])
-        else:
-            return(None)
+        return self.sorted_iter()
 
     def uris(self):
         """Extract sorted list of URIs for resources in this ResourceListDict."""
-        return(sorted(self.keys()))
+        return sorted(self.keys())
+
+    def add(self, resource, replace=False):
+        """Add just a single resource."""
+        uri = resource.uri
+        if (uri in self.keys() and not replace):
+            raise ResourceListDupeError(
+                "Attempt to add resource already in resource_list")
+        self[uri] = resource
+
+
+class ResourceListOrdered(OrderedDict, SortedIterMixin):
+    """Alternative implementation of class to store resources in ResourceList.
+
+    Key properties of this class are:
+    - has add(resource) method
+    - is iterable and gives resources (not keys) in the order added
+    """
+
+    def __iter__(self):
+        """Iterator over the resources in the list in the order added."""
+        return iter(self.values())
+
+    def uris(self):
+        """Extract list of all resource URIs in the order added."""
+        return self.keys()
 
     def add(self, resource, replace=False):
         """Add just a single resource."""
@@ -60,42 +94,6 @@ class ResourceListDict(dict):
             raise ResourceListDupeError(
                 "Attempt to add resource already in resource_list")
         self[uri] = resource
-
-
-class ResourceListOrdered(list):
-    """Alternative implementation of class to store resources in ResourceList.
-
-    FIXME - This is a rather inefficient implementation which involves
-    scanning all resources to check for duplicates. Designed just to enable
-    re-creation of examples in the spec. Something dictionary based would
-    likely be better. Might be best to use OrderedDict but that is available
-    natively only in python >= 2.7 and this library is designed for 2.6,2.7.
-
-    Key properties of this class are:
-    - has add(resource) method
-    - is iterable and results given in order added (not the usual one!)
-    """
-
-    def uris(self):
-        """Extract list of all resource URIs (in the order added)."""
-        uris = []
-        for r in self:
-            uris.append(r.uri)
-        return(uris)
-
-    def add(self, resource, replace=False):
-        """Add a single resource, check for dupes."""
-        uri = resource.uri
-        for r in self:
-            if (uri == r.uri):
-                if (replace):
-                    r = resource
-                    return
-                else:
-                    raise ResourceListDupeError(
-                        "Attempt to add resource already in resource_list")
-        # didn't find it in list, add to end
-        self.append(resource)
 
 
 class ResourceListDupeError(Exception):
@@ -123,10 +121,12 @@ class ResourceList(ListBaseWithIndex):
 
     The default storage is unordered but the iterator imposes a canonical
     order which is alphabetical by URI. If it is desired to have
-    resources listed in the order they are added then the ResourceDictOrdered
+    resources listed in the order they are added then the ResourceListOrdered
     class may be specified on creation:
 
         rl = ResourceList( resources_class=ResourceDictOrdered )
+
+    Use of ResourceListOrdered will be faster than the default.
 
     In normal use it is expected that any Resource List Index will be
     created automatically when writing out a large Resource List in
@@ -176,11 +176,11 @@ class ResourceList(ListBaseWithIndex):
         written to work for any objects in self and sc, provided that the
         == operator can be used to compare them.
 
-        The functioning of this method depends on the iterators for self and
-        src providing access to the resource objects in URI order.
+        The functioning of this method depends on the sorted_iter() iterators
+        for self and src providing access to the resource objects in URI order.
         """
-        dst_iter = iter(self.resources)
-        src_iter = iter(src.resources)
+        dst_iter = self.resources.sorted_iter()
+        src_iter = src.resources.sorted_iter()
         same = ResourceList()
         updated = ResourceList()
         deleted = ResourceList()
@@ -189,21 +189,19 @@ class ResourceList(ListBaseWithIndex):
         src_cur = next(src_iter, None)
         while ((dst_cur is not None) and (src_cur is not None)):
             # print 'dst='+dst_cur+'  src='+src_cur
-            if (dst_cur.uri == src_cur.uri):
+            if dst_cur.uri == src_cur.uri:
                 if (dst_cur == src_cur):
                     same.add(dst_cur)
                 else:
                     updated.add(src_cur)
                 dst_cur = next(dst_iter, None)
                 src_cur = next(src_iter, None)
-            elif (not src_cur or dst_cur.uri < src_cur.uri):
+            elif dst_cur.uri < src_cur.uri:
                 deleted.add(dst_cur)
                 dst_cur = next(dst_iter, None)
-            elif (not dst_cur or dst_cur.uri > src_cur.uri):
+            else:  # dst_cur.uri > src_cur.uri:
                 created.add(src_cur)
                 src_cur = next(src_iter, None)
-            else:
-                raise Exception("this should not be possible")
         # what do we have leftover in src or dst lists?
         while (dst_cur is not None):
             deleted.add(dst_cur)