+++ tests/test_resource_container.py
@@ -1,6 +1,8 @@
+"""Tests for resync.resource_container."""
+
 import unittest
 from resync.resource import Resource
-from resync.resource_container import ResourceContainer
+from resync.resource_container import ResourceContainer, _str_datetime_now
 
 
 class TestResourceContainer(unittest.TestCase):
@@ -8,9 +10,12 @@ class TestResourceContainer(unittest.TestCase):
     def test01_create_and_add(self):
         rc = ResourceContainer(resources=[])
         self.assertEqual(len(rc.resources), 0, "empty")
-        rc.resources.append(Resource('a', timestamp=1))
-        rc.resources.append(Resource('b', timestamp=2))
+        rc.add(Resource('a', timestamp=1))
+        rc.add(Resource('b', timestamp=2))
         self.assertEqual(len(rc.resources), 2, "two resources")
+        # Add two more
+        rc.add([Resource('c'), Resource('d')])
+        self.assertEqual(len(rc.resources), 4, "four resources")
 
     def test02_iter(self):
         rc = ResourceContainer(resources=[])
@@ -93,3 +98,37 @@ class TestResourceContainer(unittest.TestCase):
         self.assertEqual(rc.up, None)
         rc.up = "up_uri"
         self.assertEqual(rc.up, "up_uri")
+
+    def test09_capability(self):
+        """Test capability property."""
+        rc = ResourceContainer()
+        self.assertEqual(rc.capability, None)
+        rc.capability = "fixit"
+        self.assertEqual(rc.capability, 'fixit')
+
+    def test10_link_set(self):
+        """Test link_set method."""
+        rc = ResourceContainer()
+        self.assertEqual(rc.link('alink'), None)  # non-spec rel should be supported
+        rc.link_set('alink', 'uri:l', extra1='one', extra2='two')
+        ln = rc.link('alink')
+        self.assertEqual(rc.link('alink'), {'rel': 'alink', 'href': 'uri:l',
+                                            'extra1': 'one', 'extra2': 'two'})
+        # Can update existing link
+        rc.link_set('alink', 'uri:new', extra1='1')
+        self.assertEqual(rc.link('alink'), {'rel': 'alink', 'href': 'uri:new',
+                                            'extra1': '1', 'extra2': 'two'})
+
+    def test11_index(self):
+        """Test index property."""
+        rc = ResourceContainer()
+        self.assertEqual(rc.index, None)
+        rc.index = 'uri:i'
+        self.assertEqual(rc.index, 'uri:i')
+
+    def test12_str_datetime_now(self):
+        """Test _str_datetime_now function."""
+        self.assertTrue(isinstance(_str_datetime_now(), str))
+        self.assertTrue(isinstance(_str_datetime_now('now'), str))
+        self.assertEqual(_str_datetime_now(0), '1970-01-01T00:00:00Z')
+        self.assertEqual(_str_datetime_now('2020-12-24T16:01:00Z'), '2020-12-24T16:01:00Z')