+++ tests/testlib/testcase_with_tmpdir.py
@@ -36,10 +36,4 @@ class TestCase(unittest.TestCase):
     @property
     def tmpdir(self):
         """Read-only access to _tmpdir, just in case... The rmtree scares me."""
-        #
-        # FIXME - Hack to work on python2.6 where setUpClass is not called, will
-        # FIXME - not have proper tidy as tearDownClass will not be called.
-        # FIXME - Remove when 2.6 no longer supported
-        if (not self._tmpdir and sys.version_info < (2, 7)):
-            self.setUpClass()
         return(self._tmpdir)